removed unused websocket library fixed hdop value made global myGPS variable added url /stats to see gps information

+++ lib/GPX/GPX.h
@@ -9,7 +9,7 @@ class GPX {
     void reuseLastGpxOrCreateNew(const char* currentDateTime);
     void createNewGpxFile(const char* currentDateTime);
     void writeEndOfFile();
-    const char* createNewTrackPoint(double latitude, double longitude, const char* time, const char* fix, int hdop, double speed);
+    const char* createNewTrackPoint(double latitude, double longitude, const char* time, const char* fix, double hdop, double speed);
     void appendCurrentGpxFile(const char* str);
 };
 #endif  // __GPX_H__
